requirement.txt
# Core
gradio>=4.0.0
python-dotenv>=1.0.0

# Google Gemini API
google-genai>=0.3.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
openpyxl>=3.1.0  # Đọc file Excel

# ML/Similarity
scikit-learn>=1.3.0

# Optional tùy theo model embedding nếu muốn tìm riêng.
# sentence-transformers>=2.2.0